Update link URLs on the home page (#4)
* Update Get started link under "Lightweight"

I think it makes sense for the Get Started link under the "Light weight" section to point to the getting started page instead of the Github repo page. This make the behavior consistent with what the other Get Started links on the page do.

* Update Browse Modules link

Updating the Browse Modules link to point to the CBT.Modules project.